+++ cmd/forklift/dev-env.go
@@ -25,7 +25,7 @@ func devEnvShowAction(c *cli.Context) error {
 	return printEnvInfo(envPath)
 }
 
-// cache
+// cache-repo
 
 func devEnvCacheRepoAction(c *cli.Context) error {
 	envPath, err := dev.FindParentEnv(c.String("cwd"))
@@ -47,14 +47,35 @@ func devEnvCacheRepoAction(c *cli.Context) error {
 	// TODO: download all Docker images used by packages in the repo - either by inspecting the
 	// Docker stack definitions or by allowing packages to list Docker images used.
 	fmt.Println(
-		// TODO: add a command to check if resource constraints are all satisfied
-		// "Done! Next, you might want to run `forklift dev env check` or `forklift dev env apply`.",
-		"Done! Next, you might want to run `forklift dev env apply`.",
+		// TODO: add a command to check if resource constraints are all satisfied, and show the command
+		// to run for that (`forklift dev env check`)
+		"Done! Next, you might want to run `sudo -E forklift dev env apply`.",
 	)
 	return nil
 }
 
-// deploy
+// cache-img
+
+func devEnvCacheImgAction(c *cli.Context) error {
+	envPath, err := dev.FindParentEnv(c.String("cwd"))
+	if err != nil {
+		return errors.Wrap(err, "The current working directory is not part of a Forklift environment.")
+	}
+	wpath := c.String("workspace")
+	if !workspace.Exists(workspace.CachePath(wpath)) {
+		return errMissingCache
+	}
+
+	fmt.Println("Downloading Docker container images specified by the development environment...")
+	if err := downloadImages(envPath, workspace.CachePath(wpath)); err != nil {
+		return err
+	}
+	fmt.Println()
+	fmt.Println("Done! Next, you'll probably want to run `sudo -E forklift dev env apply`.")
+	return nil
+}
+
+// apply
 
 func devEnvApplyAction(c *cli.Context) error {
 	envPath, err := dev.FindParentEnv(c.String("cwd"))
@@ -63,8 +84,7 @@ func devEnvApplyAction(c *cli.Context) error {
 	}
 	wpath := c.String("workspace")
 	if !workspace.Exists(workspace.CachePath(wpath)) {
-		fmt.Println("The cache is empty, please run `forklift dev env cache` first")
-		return nil
+		return errMissingCache
 	}
 
 	if err := applyEnv(envPath, workspace.CachePath(wpath)); err != nil {
@@ -94,8 +114,7 @@ func devEnvShowRepoAction(c *cli.Context) error {
 	}
 	wpath := c.String("workspace")
 	if !workspace.Exists(workspace.CachePath(wpath)) {
-		fmt.Println("The cache is empty, please run `forklift dev env cache` first")
-		return nil
+		return errMissingCache
 	}
 
 	repoPath := c.Args().First()
@@ -111,8 +130,7 @@ func devEnvLsPkgAction(c *cli.Context) error {
 	}
 	wpath := c.String("workspace")
 	if !workspace.Exists(workspace.CachePath(wpath)) {
-		fmt.Println("The cache is empty, please run `forklift dev env cache` first")
-		return nil
+		return errMissingCache
 	}
 
 	return printEnvPkgs(envPath, workspace.CachePath(wpath))
@@ -127,8 +145,7 @@ func devEnvShowPkgAction(c *cli.Context) error {
 	}
 	wpath := c.String("workspace")
 	if !workspace.Exists(workspace.CachePath(wpath)) {
-		fmt.Println("The cache is empty, please run `forklift dev env cache` first")
-		return nil
+		return errMissingCache
 	}
 
 	pkgPath := c.Args().First()
@@ -144,8 +161,7 @@ func devEnvLsDeplAction(c *cli.Context) error {
 	}
 	wpath := c.String("workspace")
 	if !workspace.Exists(workspace.CachePath(wpath)) {
-		fmt.Println("The cache is empty, please run `forklift dev env cache` first")
-		return nil
+		return errMissingCache
 	}
 
 	return printEnvDepls(envPath, workspace.CachePath(wpath))
@@ -160,8 +176,7 @@ func devEnvShowDeplAction(c *cli.Context) error {
 	}
 	wpath := c.String("workspace")
 	if !workspace.Exists(workspace.CachePath(wpath)) {
-		fmt.Println("The cache is empty, please run `forklift dev env cache` first")
-		return nil
+		return errMissingCache
 	}
 
 	deplName := c.Args().First()