+++ internal/clients/docker/images.go
@@ -4,11 +4,17 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"io"
 	"strings"
 
 	"github.com/docker/cli/cli/command/inspect"
+	"github.com/docker/cli/cli/streams"
+	"github.com/docker/cli/cli/trust"
+	"github.com/docker/distribution/reference"
 	dt "github.com/docker/docker/api/types"
 	dtf "github.com/docker/docker/api/types/filters"
+	dtr "github.com/docker/docker/api/types/registry"
+	"github.com/docker/docker/pkg/jsonmessage"
 	"github.com/pkg/errors"
 )
 
@@ -100,4 +106,73 @@ func CompareDeletedImages(i, j dt.ImageDeleteResponseItem) int {
 	case i.Deleted > j.Deleted:
 		return 1
 	}
+}
+
+func (c *Client) PullImage(
+	ctx context.Context, taggedName string, outStream *streams.Out,
+) (trust.ImageRefAndAuth, error) {
+	// This function is adapted from the github.com/docker/cli/cli/command/image
+	// package's RunPull function, which is licensed under Apache-2.0. This function was changed to
+	// assume that the name is already tagged and normalized and that no auth or content trust image
+	// verification is needed.
+	distributionRef, err := reference.ParseNormalizedNamed(taggedName)
+	switch {
+	case err != nil:
+		return trust.ImageRefAndAuth{}, err
+	case reference.IsNameOnly(distributionRef):
+		return trust.ImageRefAndAuth{}, errors.Errorf(
+			"image %s must be specified with a tag", taggedName,
+		)
+	}
+
+	imgRefAndAuth, err := trust.GetImageReferencesAndAuth(ctx, authResolver, taggedName)
+	if err != nil {
+		return trust.ImageRefAndAuth{}, errors.Wrapf(
+			err, "couldn't look up ref of image %s", taggedName,
+		)
+	}
+
+	if err = c.pullImage(ctx, imgRefAndAuth, outStream); err != nil {
+		return trust.ImageRefAndAuth{}, err
+	}
+
+	return imgRefAndAuth, nil
+}
+
+func NewOutStream(out io.Writer) *streams.Out {
+	return streams.NewOut(out)
+}
+
+func authResolver(ctx context.Context, index *dtr.IndexInfo) dtr.AuthConfig {
+	return dtr.AuthConfig{}
+}
+
+func (c *Client) pullImage(
+	ctx context.Context, imgRefAndAuth trust.ImageRefAndAuth, out *streams.Out,
+) (err error) {
+	// This function is adapted from the github.com/docker/cli/cli/command/image
+	// package's imagePullPrivileged function, which is licensed under Apache-2.0. This function was
+	// changed so that it doesn't use Docker's CLI and gives up immediately if the operation is
+	// unauthorized.
+	encodedAuth, err := dtr.EncodeAuthConfig(*imgRefAndAuth.AuthConfig())
+	if err != nil {
+		return err
+	}
+	responseBody, err := c.Client.ImagePull(
+		ctx, reference.FamiliarString(imgRefAndAuth.Reference()), dt.ImagePullOptions{
+			RegistryAuth: encodedAuth,
+		},
+	)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		if cerr := responseBody.Close(); cerr != nil {
+			if err == nil {
+				err = cerr
+			}
+		}
+	}()
+
+	return jsonmessage.DisplayJSONMessagesToStream(responseBody, out, nil)
 }